update gems, move history box, fix merging bug
Updated the Gemfile.lock since we pulled in rack_csrf, might
as well bump everything else too.

History box goes under the Save button for now... might need to
polish the UX there

There was a terrible bug where post-merge the next commit by an
editor would undo the results of the merge.  That was because
the merge would write the new content into the working tree,
but this would confuse git's index.  To remedy this we just
git reset --hard after any commits we make via grit that would
affect the master branch and therefore impact the working directory's
index.